Port to .JSON
With this port the app can use JSON objects to
load the WordPairs from. It has a new WordPairReader
class which can be initialized by either a JSON string
or an InputStream that reads from a JSON file. This makes
developing a feature that app retrieves word pairs from
remote databases (e.g. Firebase) very easy and smooth.