Fixed UI black cell bug
The problem was in setTextSize. It made sure that the
word's width is equal to the desiredWidth which has a
restraint equal to cell size. In other words, it made sure
the word is shown as largely as possible without getting widthier
than the cell size. This caused a bug that when it was passed a
short letter such as "I", because its width is relatively small,
it scaled it so much to make its width match the cell width. But
this caused its height to scale as much as occupying two or three
more cells.

With this fix, it will be constraint in both width and height. So
its height also will not surpass the height of a cell.